split out json config writer
