+++ ccd-config-generator/src/main/java/uk/gov/hmcts/ccd/sdk/MultiCaseTypeGenerator.java
@@ -1,15 +1,17 @@
 package uk.gov.hmcts.ccd.sdk;
 
+import com.google.common.collect.Lists;
 import com.google.common.collect.Multimap;
 import com.google.common.collect.Multimaps;
 import java.io.File;
 import java.util.List;
 import net.jodah.typetools.TypeResolver;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Component;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
 import uk.gov.hmcts.ccd.sdk.api.CCDConfig;
 
-@Component
+@Configuration
 class MultiCaseTypeGenerator {
   private final List<CCDConfig<?, ?, ?>> configs;
 
@@ -18,15 +20,25 @@ class MultiCaseTypeGenerator {
     this.configs = configs;
   }
 
-  public void generateCaseTypes(File outFolder) {
+  @Bean
+  public List<ResolvedCCDConfig<?, ?, ?>> loadConfigs() {
     Multimap<Class<?>, CCDConfig<?, ?, ?>>
         configsByDataClass = Multimaps
         .index(configs, x -> TypeResolver.resolveRawArguments(CCDConfig.class, x.getClass())[0]);
+
+    List<ResolvedCCDConfig<?, ?, ?>> result = Lists.newArrayList();
     for (Class<?> c : configsByDataClass.keySet()) {
-      ConfigGenerator generator = new ConfigGenerator(configsByDataClass.get(c));
-      File f = new File(outFolder, c.getSimpleName());
+      ConfigResolver generator = new ConfigResolver(configsByDataClass.get(c));
+      result.add(generator.resolveCCDConfig());
+    }
+    return result;
+  }
+
+  public void generateCaseTypes(File outFolder) {
+    for (ResolvedCCDConfig<?, ?, ?> c : loadConfigs()) {
+      File f = new File(outFolder, c.caseType);
       f.mkdirs();
-      generator.resolveConfig(f);
+      new JSONConfigWriter().writeConfig(f, c);
     }
   }
