+++ ccd-config-generator/src/main/resources/META-INF/spring.factories
@@ -1,2 +1,2 @@
 org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
-  uk.gov.hmcts.ccd.sdk.ConfigGenerator
+  uk.gov.hmcts.ccd.sdk.MultiCaseTypeGenerator
