+++ ccd-config-generator/src/test/java/uk/gov/hmcts/ccd/sdk/CallbackControllerTest.java
@@ -23,6 +23,7 @@ import org.springframework.test.web.servlet.ResultMatcher;
 import uk.gov.hmcts.reform.ccd.client.model.AboutToStartOrSubmitCallbackResponse;
 import uk.gov.hmcts.reform.ccd.client.model.CallbackRequest;
 import uk.gov.hmcts.reform.ccd.client.model.CaseDetails;
+import uk.gov.hmcts.reform.fpl.BulkCaseConfig;
 import uk.gov.hmcts.reform.fpl.model.CaseData;
 
 @RunWith(SpringRunner.class)
@@ -68,7 +69,7 @@ public class CallbackControllerTest {
         this.makeRequest("mid-event?page=1", "addFamilyManCaseNumber")
             .andExpect(status().isOk())
             .andReturn();
-    CaseData data = getResponseData(result);
+    CaseData data = getResponseData(result, CaseData.class);
     assertThat(data.getFamilyManCaseNumber()).isEqualTo("PLACEHOLDER");
   }
 
@@ -80,28 +81,45 @@ public class CallbackControllerTest {
   }
 
   @SneakyThrows
-  ResultActions makeRequest(String callback, String eventId) {
-    return this.mockMvc.perform(post("/callbacks/" + callback)
-        .contentType(MediaType.APPLICATION_JSON)
-        .content(new ObjectMapper().writeValueAsString(buildRequest(eventId))));
+  @Test
+  public void testEventOnDifferentCaseType() {
+    MvcResult result =
+        this.makeRequest("about-to-start", "bulk", "addFamilyManCaseNumber")
+            .andExpect(status().isOk())
+            .andReturn();
+    BulkCaseConfig.BulkCase data = getResponseData(result, BulkCaseConfig.BulkCase.class);
+    assertThat(data.getFamilyManCaseNumber()).isEqualTo("bulk-about-to-start");
   }
 
   @SneakyThrows
-  CaseData getResponseData(MvcResult result) {
+  ResultActions makeRequest(String callback, String eventId) {
+    return makeRequest(callback, "CARE_SUPERVISION_EPO", eventId);
+  }
+
+  @SneakyThrows
+  ResultActions makeRequest(String callback, String caseType, String eventId) {
+    return this.mockMvc.perform(post("/callbacks/" + callback)
+        .contentType(MediaType.APPLICATION_JSON)
+        .content(new ObjectMapper().writeValueAsString(buildRequest(caseType, eventId))));
+  }
+
+  @SneakyThrows
+  <T> T getResponseData(MvcResult result, Class<T> c) {
     ObjectMapper mapper = new ObjectMapper();
     AboutToStartOrSubmitCallbackResponse r =
         mapper.readValue(result.getResponse().getContentAsString(),
             AboutToStartOrSubmitCallbackResponse.class);
 
     String json = mapper.writeValueAsString(r.getData());
-    return mapper.readValue(json, CaseData.class);
+    return mapper.readValue(json, c);
   }
 
-  CallbackRequest buildRequest(String eventId) {
+  CallbackRequest buildRequest(String caseType, String eventId) {
     return CallbackRequest.builder()
         .eventId(eventId)
         .caseDetails(CaseDetails.builder()
             .data(Maps.newHashMap())
+            .caseTypeId(caseType)
             .build())
         .build();
   }
