+++ ccd-gradle-plugin/test-projects/java-library/src/main/java/uk/gov/hmcts/test/library/CCDConfig.java
@@ -7,6 +7,6 @@ import org.springframework.stereotype.Component;
 public class CCDConfig implements uk.gov.hmcts.ccd.sdk.api.CCDConfig<CaseData, State, Role> {
   @Override
   public void configure(ConfigBuilder<CaseData, State, Role> builder) {
-
+    builder.caseType("test", "test", "test");
   }
 }
